Abstracted IMAGE_CONFIG constants into new imageConstants file
