+++ src/constants/index.ts
@@ -1,4 +1,5 @@
+import { IMAGE_CONFIG } from './imageConstants';
 import { PATH_NAMES, NAV_CONFIG } from './routerConstants';
 import { SOCIAL_LINKS } from './socialsConstants';
 
-export { PATH_NAMES, NAV_CONFIG, SOCIAL_LINKS };
+export { IMAGE_CONFIG, PATH_NAMES, NAV_CONFIG, SOCIAL_LINKS };
